New Changes to the SRS
Changes:
1) Added a table of contents
2) Finished Non-Functional Requirements and Open Issues
3) Better Formatted

All that's left is the functional requirements. If you guys could work on that portion (in either word or latex) and push the changes, I can format it correctly tomorrow or monday. Also if you can read it and make edits where you feel neccesary that would be great.